// bookme/src/app/company/[companyName]/order/_components/(publicItems)/_OrderPageTypes/types.tsx
export type employeeType = {
    _id: string,
    employeeName: string,
    description: string,
    duration: number,
    profileImage: string, availability: boolean,
    startTime: string, endTime: string
    lunchTimeStart: string, lunchTimeEnd: string
    companyId: string
    bookings: string[]
}
export type CompanyType = {
    _id: string,
    workingHours: Record<string, { closed: boolean }>
    companyName: string
    address: string
    companyLogo: string
    phoneNumber: number
    description: string
    companyImages: string[]
    employees: [employeeType]
    bookings: string[]
}
export type OrderNavpropsType = {
    isStage: string;
    setIsStage: (stage: string) => void;
    title: string;
    Stages: string[];
    setSelectedTime: (time: Date | null) => void
};
export type StagOneProps = {
    company: CompanyType;
    isSelectEmployee: string | string[];
    setIsSelectEmployee: (employee: string) => void;
    selectedEmployeeImf: string | undefined;
    setSelectedEmployeeImf: (employeeId: string) => void;
};